Replaced the up/down speeds with the internet ESSID (name)
This was done because polling the internet speed is very
bad if you are trying to do anything on the internet, as
it uses almost all bandwidth, which causes far more problems
than it is worth.

Also adjusted the position of the revealer to be before
instead of after the internet icon.