Improvement of keystate.sh
The script now searches for all capslock and numlock states instead
of a predefined set.
In the process of this the code has also been cleaned up, to be
more concise and readable.

Additionally the scripts have all been renamed to now end with .sh
and poll.yuck updated accordingly.